Notify when he max_seconds_unlock is too low
If max_seconds_unlock (Seconds to wait before unlock) is lower than
the time it takes for a task to be processed, the task will be
considered as failed before being completed. Thus, the task will
never be completed and database conflicts will occur each time the
system retries that task until is finally discarded (removed from
the queue).

max_seconds_unlock acts as a safe-guard, but reducing its value does
not make the system to behave faster.