Fix client's queue tasks in "queued" status are not updated when "running"
Each zeo client keeps a local pool of tasks that regularly syncs with the
queue server. The status of tasks from the local pool that are in "queued"
status are never updated, even if the real status of the task in the server
pool is "running".

This might lead to some problems with transitions when the client acts as a
consumer too, cause there is a generic guard that prevents a task in "queued"
to be transitioned. Since the client "thinks" the task is "queued" (instead
of "running"), the guard returns False and the transition cannot take place.

This commit ensures that during the synchronization process (diff route), the
server always return "running" tasks, even if client has them already on it's
own local pool. On synchronization, client always drop local tasks that have
returned by the server first. Therefore, the result is that the local task
gets updated to "running" status.