Support "receive" transition from samples' main listing
This commit enables queue support for the transition "receive" when triggered
from the main samples listing. Samples to be received are added to the queue
and the transition is done asynchronously by consumers.

This transition was not supported by queue by default, so receiving a lot of
samples at same time on instances with high load, ended up freezing the instance
and samples weren't transitioned at all because of transaction commit conflicts